+++ packages/plugins/Fullscreen/src/components/Fullscreen.vue
@@ -1,8 +1,8 @@
 <template>
   <v-tooltip
-    class="polar-fullscreen-wrap"
     :left="!isHorizontal"
     :bottom="isHorizontal"
+    :disabled="hasSmallDisplay"
   >
     <template #activator="{ on, attrs }">
       <v-btn
@@ -40,7 +40,7 @@ export default Vue.extend({
     },
   },
   computed: {
-    ...mapGetters(['map']),
+    ...mapGetters(['hasSmallDisplay', 'map']),
     ...mapGetters('plugin/fullscreen', [
       'isInFullscreen',
       'renderType',
@@ -114,9 +114,4 @@ export default Vue.extend({
   },
 })
 </script>
-<style lang="scss" scoped>
-.polar-fullscreen-wrap {
-  display: flex;
-  flex-direction: column;
-}
-</style>
+<style lang="scss" scoped></style>