+++ packages/plugins/IconMenu/src/language.ts
@@ -9,6 +9,7 @@ const language: LanguageOption[] = [
           hints: {
             attributions: 'Quellennachweis',
             draw: 'Zeichenwerkzeuge',
+            filter: 'Filter',
             layerChooser: 'Kartenauswahl',
             gfi: 'Objektliste',
           },
@@ -24,6 +25,7 @@ const language: LanguageOption[] = [
           hints: {
             attributions: 'Attributions',
             draw: 'Draw tools',
+            filter: 'Filter',
             layerChooser: 'Choose map',
             gfi: 'Feature list',
           },